amazon_compare_prices/requirements.txt
amqp==5.3.1
asgiref==3.8.1
asttokens==3.0.0
beautifulsoup4==4.12.3
billiard==4.2.1
bs4==0.0.2
cachetools==5.5.0
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
confuse==2.0.1
debugpy==1.8.11
decorator==5.1.1
Django==5.1.4
django-extensions==3.2.3
django-redis==5.4.0
django-shell-plus==1.1.7
executing==2.1.0
flower==2.0.1
humanize==4.11.0
idna==3.10
ipdb==0.13.13
ipython==8.31.0
jedi==0.19.2
kombu==5.4.2
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-amazon-sp-api==1.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
